# taskboard/public/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TaskBoard API",
    "version": "1.0.0",
    "description": "A comprehensive API for managing Kanban tasks with integrated MCP (Model Context Protocol) chat functionality.\n\n## Authentication\n\nProtected endpoints require user identification via one of these methods:\n1. **Bearer Token**: `Authorization: Bearer <userId>`\n2. **Query Parameter**: `?userId=<userId>`  \n3. **Request Body**: Include `userId` field in JSON body\n\nThis flexible authentication supports both web applications and MCP integrations.\n\n## Features\n\n- User registration and authentication\n- CRUD operations for Kanban items (todo, doing, done)  \n- MCP-powered AI chat assistant integration\n- Session-based conversation management\n- Tool usage tracking and display"
  },
  "servers": [
    {
      "url": "${API_BASE_URL:-http://localhost:3000}",
      "description": "API server (configurable via API_BASE_URL environment variable)"
    }
  ],
...
  "paths": {
...
    "/api/items/{id}": {
      "get": {
        "summary": "Get specific item",
        "description": "Retrieve a specific Kanban item by ID",
        "operationId": "get_item_by_id",
        "tags": [
          "Items"
        ],
        "x-gram": {
          "name": "get_task_by_id",
          "description": "<context>\nThis endpoint retrieves a specific Kanban task by its unique ID.\nOnly tasks belonging to the authenticated user can be accessed.\n</context>\n<prerequisites>\n- userId parameter is required for authentication\n- Valid task ID must be provided in the URL path\n</prerequisites>\n",
          "responseFilterType": "jq"
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Unique identifier of the item",
            "schema": {
              "type": "string"
            },
            "example": "clh7x1234abcd5678efgh"
          },
          {
            "name": "userId",
            "in": "query",
            "description": "User ID (required for authentication)",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "clh7x1234abcd5678user"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                },
                "example": {
                  "id": "clh7x1234abcd5678efgh",
                  "userId": "clh7x1234abcd5678user",
                  "title": "Complete project documentation",
                  "details": "Write comprehensive API documentation and user guides",
                  "status": "todo",
                  "createdAt": "2025-07-31T20:30:00.000Z"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Unauthorized"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Item not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Internal server error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update item",
        "description": "Update an existing Kanban item",
        "operationId": "update_item",
        "tags": [
          "Items"
        ],
        "x-gram": {
          "name": "update_task",
          "description": "<context>\nThis endpoint updates an existing Kanban task. You can modify the title, details, status, or any combination of these fields.\nOnly tasks belonging to the authenticated user can be updated.\n</context>\n<prerequisites>\n- userId parameter is required for authentication\n- Valid task ID must be provided in the URL path\n- At least one field (title, details, or status) should be provided for update\n</prerequisites>\n<usage>\n- Use to move tasks between columns (todo → doing → done)\n- Use to update task details as work progresses\n- Use to correct typos or add more information to existing tasks\n</usage>\n",
          "responseFilterType": "jq"
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Unique identifier of the item",
            "schema": {
              "type": "string"
            },
            "example": "clh7x1234abcd5678efgh"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemRequest"
              },
              "example": {
                "title": "Updated task title",
                "details": "Updated task details with more information",
                "status": "doing",
                "userId": "clh7x1234abcd5678user"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                },
                "example": {
                  "id": "clh7x1234abcd5678efgh",
                  "userId": "clh7x1234abcd5678user",
                  "title": "Updated task title",
                  "details": "Updated task details with more information",
                  "status": "doing",
                  "createdAt": "2025-07-31T20:30:00.000Z"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Invalid status value"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Unauthorized"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Item not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Internal server error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete item",
        "description": "Delete a Kanban item",
        "operationId": "delete_item",
        "tags": [
          "Items"
        ],
        "x-gram": {
          "name": "delete_task",
          "description": "<context>\nThis endpoint permanently deletes a Kanban task from the user's board.\nThis action cannot be undone, so use with caution.\n</context>\n<prerequisites>\n- userId parameter is required for authentication\n- Valid task ID must be provided in the URL path\n</prerequisites>\n<usage>\n- Use to remove completed tasks that are no longer needed\n- Use to clean up duplicate or incorrect tasks\n- Consider moving tasks to 'done' status instead of deleting if you want to keep a record\n</usage>\n",
          "responseFilterType": "jq"
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Unique identifier of the item",
            "schema": {
              "type": "string"
            },
            "example": "clh7x1234abcd5678efgh"
          },
          {
            "name": "userId",
            "in": "query",
            "description": "User ID (required for authentication)",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "clh7x1234abcd5678user"
          }
        ],
        "responses": {
          "204": {
            "description": "Item successfully deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Unauthorized"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Item not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Internal server error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ItemStatus": {
        "type": "string",
        "enum": [
          "todo",
          "doing",
          "done"
        ],
        "description": "Current status of the item in the Kanban board"
      },
      "Item": {
        "type": "object",
        "required": [
          "id",
          "userId",
          "title",
          "details",
          "status",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the item"
          },
          "userId": {
            "type": "string",
            "description": "ID of the user who owns this item"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "description": "Title of the item"
          },
          "details": {
            "type": "string",
            "maxLength": 1000,
            "description": "Detailed description"
          },
          "status": {
            "$ref": "#/components/schemas/ItemStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UpdateItemRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200
          },
          "details": {
            "type": "string",
            "maxLength": 1000
          },
          "status": {
            "$ref": "#/components/schemas/ItemStatus"
          },
          "userId": {
            "type": "string",
            "description": "User ID (optional if provided via Bearer token or query param)"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message"
          }
        }
      }
    },
    "securitySchemes": {
      "UserIdParam": {
        "type": "apiKey",
        "in": "query",
        "name": "userId",
        "description": "User ID parameter for authentication"
      }
    }
  }
}